Allow "search_interval" to be configured
The default value is half of the smaller orbital period of either origin or destination. However a smaller value gives better results for some transfers with a large difference in the orbital period. For example, LKO to Minmus with default settings averages about 1,140 m/s. With finer grained settings it results in the true minimum of around 930 m/s.