Remove ".ks" extension from source filenames and add "compile_to" function
Allow users to use the more compact compiled version of the source on craft that have limited volume space.
By removing the extension kOS will first look for files with a ".ksm" extension, then for files with a ".ks" extension.
Compiled files take about 20% of the space of the raw source, resulting in a significant space savings.

Refactor the "import" method to use a list to support a "compileto" function that takes the same list of source files.
This new convenience function compiles the source to a user-specified destination (e.g. a kOS processor hard drive).