New mathematical approach to refine intercepts
Refine Vessel-to-body and Body-to-body transfers based on the approach in the paper:
"A new method of patched-conic for interplanetary orbit" by Jin Li, Jianhui Zhao and Fan Li
https://doi.org/10.1016/j.ijleo.2017.10.153

Given a velocity vector at SOI boundary, periapsis altitude and inclination, calculate the position vector "r" that satisifies these contraints. This position vector is combined in a feedback loop with the Lambert solver to refine the initial estimate from one that only considers planets as points to one that takes the SOI spheres into account. This removes the need both for the "impact_parameter" function and the line search algorithm simplifying the vessel-to-body case. The body-to-body case also simplifies into a single step instead of two.

As a breaking change, "polar" orbit orientations are no longer supported. The theoretical math is possible, however the actual implementation results in poor intercepts that frequently collide with the body. Retrograde orbits are still supported.

[orbit.ks]
* Remove "impact_parameter" function
* Add "duration_from_soi_edge" and "offset_from_soi_edge" functions
* Minor comment and code cleanup

[transfer.ks]
* Streamline "vessel_to_body" function to use new approach
* Streamline "body_to_body" function to use new approach

[validate.ks]
* Remove "polar" as a valid option for "final_orbit_orientation"

[README.md]
* Remove polar details from "final_orbit_orientation" description
* Remove high delta-v known issue that only applies to polar orbits