Fix periapsis time calculation when intercept orbit is elliptical
Hyperbolic intercept orbits always have a negative mean anomaly at epoch, so calculating the time duration to periapsis (when mean anomaly is zero) is always correct.
Elliptical orbits have a positive mean anomaly at epoch, that meant the time at *previous* periapsis in the past was being returned.

Add a simple check so that if the mean anomaly difference is less than zero, bump it by 2π (one whole orbit) so that the time to *next* periapsis is returned.