Take refinement into account before returning total deltav for transfer
Create new "refine.ks" script that composes the deltav "cost" function used by "search.ks".

For vessels and planets with a SOI to Periapsis ratio of less than 1% the existing strategy is used. This strategy considers the planets as a zero-width point on the assumption that over large distance only small adjustments will be need to the predicted transfer.

However some bodies in KSP have extremely large SOI to Periapsis ratios, for example Ike 38%, Mun 23% and Tylo 17%. For transfers to and from these bodies, the SOI sphere needs to be taken into account when finding the lowest cost transfer. The "offset_from_soi_edge" and "duration_from_soi_edge" function in "orbit.ks" predict the time and position where a craft will exit/enter SOI.

Interestingly another challenge with very large SOI to Periapsis ratios is that the minimum escape velocity can be higher than the desired transfer. For example a direct Hohmann transfer from Laythe to Vall is impossible, as the minmum escape velocity from Laythe is *higher* than the velocity required. The search algorithm now supports a minimum value for deltav.

[refine.ks]
* Extract composition of search function

[orbit.ks]
* Add "minimum_escape_velocity" function
* Re-order "time_at_soi_edge" function to more logical location

[search.ks]
* Remove delta-v cost function
* Add typesafe rounding function

[transfer.ks]
* Convert accidentally global function to local