Restore polar option to final orbit orientation
For polar orbits either +90 or -90 inclination will do. This approach chooses +90 as slightly more traditional. The simplified formula for polar orbit is slightly different to prograde/retrograde as the cosine of the inclination is zero. However I've no idea why the sign of the x and z components must be opposite!

Restore option to "validate.ks" and "README.md"