Add validation check for minimum insertion velocity
Add a safety check that the predicted insertion velocity at the boundary of the destination's SOI is above the minimum possible (based on the desired final periapsis.

This is similar to the existing check that the ejection velocity from the origin is greater than the minimum possible.

This fixes an error occuring in some scenarios when the initial raw point-to-point Lamber solution predicted an intercept velocity less than this minimum.